requirements.txt
pandas
requests
tweepy
urllib3
bs4
mysqlclient
pyforms